+++ src/main/java/io/github/codesmashing/programming_project_aquafin_case2/programming_project_aquafin_case2/models/PrecipitationData.java
@@ -2,6 +2,8 @@ package io.github.codesmashing.programming_project_aquafin_case2.programming_pro
 
 import java.time.LocalDate;
 
+import org.springframework.format.annotation.DateTimeFormat;
+
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
 import jakarta.persistence.*;
@@ -26,6 +28,7 @@ public class PrecipitationData {
 
 	@NotNull
 	@Column(nullable = false)
+	@DateTimeFormat(pattern = "dd/MM/yyyy")
 	private LocalDate dateOccurence;
 
 	@PositiveOrZero